move svg icon file
